+++ docs/TehPers.FishingOverhaul/CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog for [Teh's Fishing Overhaul].
 - Add priority tiers to `AvailabilityInfo`. When selecting a fish/trash/treasure, only the entries
   that have the highest priority tier from the set of available entries will be considered.
 - Add a `tfo_export` console command to export all registered fish/trash/treasure entries to a file.
+- Add an option to show trash entries in the fishing HUD.
 
 ### Changed